# shop-service/openapi.yaml
openapi: 3.0.3
info:
  title: Historical Computer Parts Shop Service API
  version: 0.2.0
  description: External shop, checkout and protected admin API for historical computer parts.
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /products:
    get:
      summary: List active products
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /orders:
    post:
      summary: Create an order
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "202":
          description: Order accepted for asynchronous processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "422":
          description: Validation or business rule failure
  /orders/{orderId}:
    get:
      summary: Get order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current order state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: Order not found
  /admin/login:
    post:
      summary: Login as admin and receive an HttpOnly session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login accepted
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
  /admin/logout:
    post:
      summary: Logout admin session
      responses:
        "200":
          description: Session cleared
  /admin/session:
    get:
      summary: Check current admin session
      security:
        - AdminCookie: []
      responses:
        "200":
          description: Active admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Admin login required
  /admin/orders:
    get:
      summary: List recent orders for the admin monitor
      security:
        - AdminCookie: []
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrder"
        "401":
          description: Admin login required
  /admin/orders/{orderId}/audit:
    get:
      summary: Get protected audit timeline for an order
      security:
        - AdminCookie: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Audit timeline
        "401":
          description: Admin login required
        "404":
          description: Order not found
components:
  securitySchemes:
    AdminCookie:
      type: apiKey
      in: cookie
      name: admin_session
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
  schemas:
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: shop-service
    Product:
      type: object
      required: [id, name, year, description, price, currency, imageUrl, imageAlt, imageSource, imageLicense, imageCredit]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        year:
          type: string
        description:
          type: string
        price:
          type: string
        currency:
          type: string
        imageUrl:
          type: string
          format: uri
        imageAlt:
          type: string
        imageSource:
          type: string
          format: uri
        imageLicense:
          type: string
        imageCredit:
          type: string
    CreateOrderRequest:
      type: object
      required: [customer, shippingAddress, items, payment]
      properties:
        customerId:
          type: string
          format: uuid
        customer:
          $ref: "#/components/schemas/Customer"
        shippingAddress:
          $ref: "#/components/schemas/Address"
        billingAddress:
          $ref: "#/components/schemas/Address"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        payment:
          $ref: "#/components/schemas/PaymentSelection"
    Customer:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
    Address:
      type: object
      required: [street, houseNumber, postalCode, city, country]
      properties:
        street:
          type: string
        houseNumber:
          type: string
        postalCode:
          type: string
        city:
          type: string
        country:
          type: string
    OrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    PaymentSelection:
      type: object
      required: [provider, currency]
      properties:
        provider:
          type: string
          enum: [paypal, stripe]
        currency:
          type: string
          example: EUR
        scenario:
          type: string
          description: Internal test scenario; omitted by the customer frontend.
          default: happy_path
        mode:
          type: string
          enum: [sandbox]
          default: sandbox
        cardholder:
          type: string
          description: Non-sensitive display metadata for Stripe sandbox checkout.
        testPaymentMethod:
          type: string
          description: Stripe sandbox payment method id, for example pm_card_visa.
        paypalEmail:
          type: string
          description: PayPal sandbox buyer email entered in the checkout UI.
        paypalOrderId:
          type: string
          description: PayPal sandbox order id approved by the buyer.
        paypalCaptureId:
          type: string
          description: PayPal sandbox capture id after successful capture.
        approveUrl:
          type: string
          description: PayPal approval URL opened by the checkout UI.
        status:
          type: string
          description: Last known PayPal sandbox status.
    OrderResponse:
      type: object
      required: [orderId, correlationId, status]
      properties:
        orderId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        status:
          type: string
        amount:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
        username:
          type: string
          nullable: true
    AdminOrder:
      type: object
      required: [orderId, correlationId, status, amount, currency, createdAt, updatedAt]
      properties:
        orderId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        status:
          type: string
        amount:
          type: string
        currency:
          type: string
        customer:
          type: object
        shippingAddress:
          type: object
        billingAddress:
          type: object
          nullable: true
        items:
          type: array
          items:
            type: object
        payment:
          type: object
          nullable: true
        transactionId:
          type: string
          nullable: true
        invoiceId:
          type: string
          format: uuid
          nullable: true
        invoiceStatus:
          type: string
          nullable: true
        warehouseCommitStatus:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time